Split analysis methods into separate utils
